pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lightcnn_pytorch"
version = "1.2.0"
description = "LightCNN for fast, accurate and lightweight face verification"
authors = [
    {name = "Yernar Bekbolat", email = "dvayernar@gmail.com"},
]

readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "numpy",
    "torch",
    "torchvision",
    "opencv-python",
    "dlib",
    "scipy",
    "pillow",
    "gdown",
]

[tool.setuptools]
packages = ["lightcnn_pytorch"]
package-data = {"lightcnn_pytorch" = ["weights/*.pth.tar", "weights/*.dat"]}

[tool.black]
line-length = 88
target-version = ['py37']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3 